requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Web scraping (for article collection)
aiohttp>=3.9.0
feedparser>=6.0.10
trafilatura>=1.6.0
beautifulsoup4>=4.12.0

# Data processing
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For advanced features
requests>=2.31.0

